Adds two lines of code to make singularity ships possible.
This is my first time using GitHub so don’t be to harsh on me.

But this makes it so uranium floor is immune to the singularity. Currently it’s impossible to build a working singularity on a moving ship because of the open space it creates. So I went with the most simple fix thanks to my lack of knowledge. I’m open of suggestions to fix/ improve this or if I should change the immune floor.