Makes singularity ships possible. (#34)
makes it so uranium floor is immune to the singularity. Currently it’s impossible to build a working singularity on a moving ship because of the open space it creates. So I went with the most simple fix thanks to my lack of knowledge. I’m open of suggestions to fix/ improve this or if I should change the immune floor.