Begins the move to modularization (#33)
I plan to setup a PR bot for Shiptest soon, to make this a proper downstream and to stay up to date.
However to prepare for this, we need to get out custom code in order. As such, I've begun to comment what has been already changed, as well as sort our maps into it's own folders.
Aswell, I added a guide on maintaining a modular setup, thanks to skyrat for the guide!