Requested Fix from Jack (#77)
Adds a couple more mobs to the ruin.
Increases damage (from 10/15 to 20/35) and health (from 100 to 200) of illusion mobs.
Fixes mobs spawning with 25 missing health.
Replaces some shrapnel mines with Stingbang, chilled vapor and stun mines (These can be triggered by the mobs in the ruin)
Removes the Theatre access that the Reticence Mech requires.

Comment: All mechs are immune to projectiles as of making this commit. This is not an issue of the ruin and therefore should not be taken into account. Ready for merging.