Adds Bluespace Miner to replace Deep-Core Drills (#197)
This adds the new Bluespace Miner to the game, the bluespace miner unlike the Deep-Core mines on your ship and aslong as it has power, to compensate for this it generates resources slower and requires a hefty amount of power. Upgrading the laser will double the materials generated with no downside. However, upgrading the capacitor will give a large boost to materials produced at the cost of an increase in power draw, you can find the perfect balance of power drain to resources generated for your ship with this.