Space Ruin - Nanotrasen Office
A ruin I made for Shiptest, modified to be compatible with this codebase.

Adds a new ruin, a semi-recently abandoned Nanotrasen office building. This is my first ruin, and now that I have familiarized myself with the tools and processes involved, I will be making larger ones in the near future.




There is a decent amount of gear in this ruin, with the main loot item being an Origami handbook and a couple handguns, as well as two security lockers full of the default loot. However, this is all defended by a couple dozen stickmen and stickdogs, as well as Mjor the Creative.

Features restorable power

Extended list of loot:
The Art of Origami
Dronespeak Manual
Random Manual x4
Commander x2
SecTech vendor
Security Officer's Locker x2
Jukebox
Personal Ai device
Paperwork the sloth
Deep Fryer x2
Microwave x2
Refrigerator (with milk and eggs) x4
Super P.A.C.M.A.N. 
Uranium sheet x5
Worn-Out Insulated Gloves
Hundreds of sheets of paper
Genius research papers (hidden)


![image](https://user-images.githubusercontent.com/19318747/187530967-887db674-8634-4a4b-bd4d-6917f4d80abf.png)

![image](https://user-images.githubusercontent.com/19318747/187531032-e98d8a2b-6b52-4e66-a3b7-7c6f62c039f3.png)

I affirm that I have tested all of my proposed changes and that any issues found during testing have been addressed.

## Why is it good for the game?

We need more ruins, especially in space, and is an interesting one to explore. There is decent loot and a high risk of death due to many enemies. I've run through it with the starter gear from a couple of ships, and while it is difficult, it isn't impossible.